add functional tests for dircolors
Actually write test cases for dircolors formatting, both the lower-level
format_mode() and the higher-level format() with actual files. To test
on real files, a temporary directory is created and populated with some
files, and it's cleaned up at the end.

test_symlink_target() is broken because of an actual bug: format was
trying to stat link's targets relative to the main cwd rather than
relative to the link itself.

Also update README to check this off the TODO list.